Fix(deps): Update docker/setup-buildx-action commit SHA (#6)
The previous commit SHA (d70bba72b6f31a22640103738a088e5d3c8b4104) was no longer valid, causing the 'build-scan-push' job in the Docker workflow to fail.

This commit updates the SHA to the latest version (v3.11.1), which resolves the issue. This also includes the fix for the `docker/setup-qemu-action`.